Switched to two-dimensional indexing
Done in order [x][y], which does not fit naturally with the way bytes come out
but does with the order PIL address tuples use and thus what was already there.
This is a bit inefficient in creating textures and outputting as Images,
but not in the central loop where it's actually important.